+++ plugins/newspack-plugin/CHANGELOG.md
@@ -1,3 +1,10 @@
+## newspack [6.45.2](https://github.com/Automattic/newspack-workspace/compare/newspack@6.45.1...newspack@6.45.2) (2026-07-14)
+
+
+### Bug Fixes
+
+* **content-gate:** keep institutional-access check URL host-relative ([#614](https://github.com/Automattic/newspack-workspace/issues/614)) ([e72f8e2](https://github.com/Automattic/newspack-workspace/commit/e72f8e2303f308ae768b504b27b5977e1b9248d0))
+
 ## newspack [6.45.1](https://github.com/Automattic/newspack-workspace/compare/newspack@6.45.0...newspack@6.45.1) (2026-07-09)